Split out data for smaller memory footprint
