+++ src/main/java/org/alliancegenome/curation_api/controllers/crud/CodingSequenceCrudController.java
@@ -5,9 +5,9 @@ import java.util.List;
 import org.alliancegenome.curation_api.controllers.base.BaseEntityCrudController;
 import org.alliancegenome.curation_api.dao.CodingSequenceDAO;
 import org.alliancegenome.curation_api.interfaces.crud.CodingSequenceCrudInterface;
-import org.alliancegenome.curation_api.jobs.executors.Gff3CDSExecutor;
-import org.alliancegenome.curation_api.jobs.executors.Gff3CDSLocationExecutor;
-import org.alliancegenome.curation_api.jobs.executors.Gff3TranscriptCDSExecutor;
+import org.alliancegenome.curation_api.jobs.executors.gff.Gff3CDSExecutor;
+import org.alliancegenome.curation_api.jobs.executors.gff.Gff3CDSLocationExecutor;
+import org.alliancegenome.curation_api.jobs.executors.gff.Gff3TranscriptCDSExecutor;
 import org.alliancegenome.curation_api.model.entities.CodingSequence;
 import org.alliancegenome.curation_api.model.entities.bulkloads.BulkLoadFileHistory;
 import org.alliancegenome.curation_api.model.ingest.dto.fms.Gff3DTO;
