+++ src/main/java/org/alliancegenome/curation_api/controllers/crud/ExonCrudController.java
@@ -5,9 +5,9 @@ import java.util.List;
 import org.alliancegenome.curation_api.controllers.base.BaseEntityCrudController;
 import org.alliancegenome.curation_api.dao.ExonDAO;
 import org.alliancegenome.curation_api.interfaces.crud.ExonCrudInterface;
-import org.alliancegenome.curation_api.jobs.executors.Gff3ExonExecutor;
-import org.alliancegenome.curation_api.jobs.executors.Gff3ExonLocationExecutor;
-import org.alliancegenome.curation_api.jobs.executors.Gff3TranscriptExonExecutor;
+import org.alliancegenome.curation_api.jobs.executors.gff.Gff3ExonExecutor;
+import org.alliancegenome.curation_api.jobs.executors.gff.Gff3ExonLocationExecutor;
+import org.alliancegenome.curation_api.jobs.executors.gff.Gff3TranscriptExonExecutor;
 import org.alliancegenome.curation_api.model.entities.Exon;
 import org.alliancegenome.curation_api.model.entities.bulkloads.BulkLoadFileHistory;
 import org.alliancegenome.curation_api.model.ingest.dto.fms.Gff3DTO;
