+++ src/main/java/org/alliancegenome/curation_api/controllers/crud/TranscriptCrudController.java
@@ -5,11 +5,11 @@ import java.util.List;
 import org.alliancegenome.curation_api.controllers.base.BaseEntityCrudController;
 import org.alliancegenome.curation_api.dao.TranscriptDAO;
 import org.alliancegenome.curation_api.interfaces.crud.TranscriptCrudInterface;
-import org.alliancegenome.curation_api.jobs.executors.Gff3TranscriptCDSExecutor;
-import org.alliancegenome.curation_api.jobs.executors.Gff3TranscriptExecutor;
-import org.alliancegenome.curation_api.jobs.executors.Gff3TranscriptExonExecutor;
-import org.alliancegenome.curation_api.jobs.executors.Gff3TranscriptGeneExecutor;
-import org.alliancegenome.curation_api.jobs.executors.Gff3TranscriptLocationExecutor;
+import org.alliancegenome.curation_api.jobs.executors.gff.Gff3TranscriptCDSExecutor;
+import org.alliancegenome.curation_api.jobs.executors.gff.Gff3TranscriptExecutor;
+import org.alliancegenome.curation_api.jobs.executors.gff.Gff3TranscriptExonExecutor;
+import org.alliancegenome.curation_api.jobs.executors.gff.Gff3TranscriptGeneExecutor;
+import org.alliancegenome.curation_api.jobs.executors.gff.Gff3TranscriptLocationExecutor;
 import org.alliancegenome.curation_api.model.entities.Transcript;
 import org.alliancegenome.curation_api.model.entities.bulkloads.BulkLoadFileHistory;
 import org.alliancegenome.curation_api.model.ingest.dto.fms.Gff3DTO;
